requirements.txt
algokit-utils==1.1.1
beaker-pyteal==1.0.0
cffi==1.15.1
docstring-parser==0.14.1
exceptiongroup==1.1.1
iniconfig==2.0.0
msgpack==1.0.5
packaging==23.1
pluggy==1.0.0
py-algorand-sdk==2.2.0
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
PyNaCl==1.5.0
pyteal==0.23.0
pytest==7.3.1
semantic-version==2.10.0
tomli==2.0.1
